refactor: update import paths for site title and description constants
