+++ visual_dashboard/src/app/dashboard/anaxa/page.tsx
@@ -1,4 +1,4 @@
-import { SITE_TITLE, SITE_DESCRIPTION } from "@/app/layout";
+import { SITE_DESCRIPTION, SITE_TITLE } from "@/app/constants";
 import { Metadata } from "next";
 
 export const metadata: Metadata = {
